completed EditAccountTest
